Tutorial Maker: use the Script Editor instead of ImageJ's Editor
The script editor is Swing-based, so it is easier to extend, and more
robust all-over.  Besides, it just gained a few nice features, such
as "remove trailing white space", and there is no reason to believe
that it will not gain more attractive features in the near future.

So let's just use that editor.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>