MediaWikiClient: allow a different session key than wikidb_session
For the upcoming change to allow to upload to different Wikis than the
Fiji Wiki, we need to be a little bit less strict in what we accept.
The common theme of MediaWiki seems to be that the session key starts
with 'wikidb_' and ends in 'session', so let's accept such session keys.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>