Wiki Editor: reuse MediaWikiClient for wikiHasImage()
Earlier, we tried to be cute and check whether an image is already there
by accessing the Wiki without logging in first.

However, that does not work well if you happen to edit a Wiki where you
are required to log in even to see the pages.

So let's reuse that code from the MediaWikiClient.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>