Tutorial Maker: offer to make a white background window
When making screenshots, it is often useful to have a totally white
background.

The new Wiki>Make white background image command does that. It opens a
new window that is undecorated and maximized, and by default it is in
the background.

The white window understands a few keyboard bindings: The Enter key
will get the ImageJ window into foreground, the Space key will put
the white window into the background, and the Escape key will close
the window.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>